diag(triton): int64 offset cast in silu/pixel_shuffle_broadcast/add_bias_broadcast — garbage persists
Phase 1 of int64 overflow hypothesis test — cast offset arithmetic
to int64 in:
- pixel_shuffle_broadcast_aware_kernel (commit c593ad7 origin)
- add_bias_broadcast_kernel (commit a8c31cc origin)
- silu_forward_kernel (broader element-wise candidate)

Both etape16 (2 new kernels) and etape17 (silu) produce IDENTICAL
green texture noise PNG output on Sana 4Kpx triton compiled. The
int32 overflow in these specific 3 kernels is NOT the root cause
of the numerical garbage at 4Kpx scale.

Anti-régression Sana 1024 triton not yet verified post-changes;
the int64 cast is safe (offset range fits in int64 trivially), so
no functional regression expected.

The garbage pattern is robustly reproduced regardless of which
element-wise kernel is touched — strongly suggests the offset
overflow is in a DIFFERENT op in the VAE post-pixel_shuffle chain
(e.g. conv2d, rms_norm, _to_copy) or the issue is not int32
overflow at all.

Next factual step (Phase 1 extended): tensor-value diff between
sequential and triton_sequential at each VAE op — identify the
first op where output values diverge, that op's kernel is the
bug source.

+++ src/neurobrix/kernels/ops/add.py
@@ -53,7 +53,11 @@ def add_bias_broadcast_kernel(
     addressing.
     """
     pid = tl.program_id(0)
-    offset = pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
+    # Cast to int64 — tensors >= 2^31 elements (e.g. Sana 4Kpx VAE
+    # add::88 input 1x4096x4096x128 = 2^31) overflow int32 offset
+    # arithmetic silently and corrupt `offset % feat_dim`, producing
+    # garbage output. P-SANA-4KPX-RUNTIME 2026-05-07.
+    offset = (pid * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)).to(tl.int64)
     mask = offset < n_elements
 
     c = offset % feat_dim
